rename logging -> logger to avoid confusion with built-in logging module
